Enum tungstenite::Message

pub enum Message {
    Text(String),
    Binary(Vec<u8>),
    Ping(Vec<u8>),
    Pong(Vec<u8>),
    Close(Option<CloseFrame<'static>>),
}

An enum representing the various forms of a WebSocket message.

Variants

Text(String)

A text WebSocket message

Tuple Fields of Text

0: String

Binary(Vec<u8>)

A binary WebSocket message

Tuple Fields of Binary

0: Vec<u8>

Ping(Vec<u8>)

A ping message with the specified payload

The payload here must have a length less than 125 bytes

Tuple Fields of Ping

0: Vec<u8>

Pong(Vec<u8>)

A pong message with the specified payload

The payload here must have a length less than 125 bytes

Tuple Fields of Pong

0: Vec<u8>

Close(Option<CloseFrame<'static>>)

A close message with the optional close frame.

Tuple Fields of Close

0: Option<CloseFrame<'static>>

Implementations

impl Message

pub fn text<S>(string: S) -> Message where
S: Into<String>,

Create a new text WebSocket message from a stringable.

pub fn binary<B>(bin: B) -> Message where
B: Into<Vec<u8>>,

Create a new binary WebSocket message by converting to Vec.

pub fn is_text(&self) -> bool

Indicates whether a message is a text message.

pub fn is_binary(&self) -> bool

Indicates whether a message is a binary message.

pub fn is_ping(&self) -> bool

Indicates whether a message is a ping message.

pub fn is_pong(&self) -> bool

Indicates whether a message is a pong message.

pub fn is_close(&self) -> bool

Indicates whether a message ia s close message.

pub fn len(&self) -> usize

Get the length of the WebSocket message.

pub fn is_empty(&self) -> bool

Returns true if the WebSocket message has no content. For example, if the other side of the connection sent an empty string.

pub fn into_data(self) -> Vec<u8>

Consume the WebSocket and return it as binary data.

pub fn into_text(self) -> Result<String>

Attempt to consume the WebSocket message and convert it to a String.

pub fn to_text(&self) -> Result<&str>

Attempt to get a &str from the WebSocket message, this will try to convert binary data to utf8.