Enum tungstenite::error::Error

pub enum Error {
    ConnectionClosed,
    AlreadyClosed,
    Io(Error),
    Capacity(Cow<'static, str>),
    Protocol(Cow<'static, str>),
    SendQueueFull(Message),
    Utf8,
    Url(Cow<'static, str>),
    Http(StatusCode),
    HttpFormat(Error),
}

Possible WebSocket errors

Variants

ConnectionClosed

WebSocket connection closed normally. This informs you of the close. It's not an error as such and nothing wrong happened.

This is returned as soon as the close handshake is finished (we have both sent and received a close frame) on the server end and as soon as the server has closed the underlying connection if this endpoint is a client.

Thus when you receive this, it is safe to drop the underlying connection.

Receiving this error means that the WebSocket object is not usable anymore and the only meaningful action with it is dropping it.

AlreadyClosed

Trying to work with already closed connection.

Trying to read or write after receiving ConnectionClosed causes this.

As opposed to ConnectionClosed, this indicates your code tries to operate on the connection when it really shouldn't anymore, so this really indicates a programmer error on your part.

Io(Error)

Input-output error. Apart from WouldBlock, these are generally errors with the underlying connection and you should probably consider them fatal.

Capacity(Cow<'static, str>)

When reading: buffer capacity exhausted.

When writing: your message is bigger than the configured max message size (64MB by default).

Protocol(Cow<'static, str>)

Protocol violation.

SendQueueFull(Message)

Message send queue full.

Utf8

UTF coding error

Url(Cow<'static, str>)

Invalid URL.

Http(StatusCode)

HTTP error.

HttpFormat(Error)

HTTP format error.