tungstenite

Function client_tls_with_config

pub fn client_tls_with_config<R, S>(
    request: R,
    stream: S,
    config: Option<WebSocketConfig>,
    connector: Option<Connector>,
) -> Result<(WebSocket<MaybeTlsStream<S>>, Response), HandshakeError<ClientHandshake<MaybeTlsStream<S>>>>
where
    R: IntoClientRequest,
    S: Read + Write,

The same as client_tls() but one can specify a websocket configuration, and an optional connector. If no connector is specified, a default one will be created.

Please refer to client_tls() for more details.