Struct tungstenite::handshake::headers::Headers

pub struct Headers { /* fields omitted */ }

HTTP request or response headers.

Methods

impl Headers

fn find_first(&self, name: &str) -> Option<&[u8]>

Get first header with the given name, if any.

fn find<'headers, 'name>(&'headers self,
name: &'name str)
-> HeadersIter<'name, 'headers>

Iterate over all headers with the given name.

fn header_is(&self, name: &str, value: &str) -> bool

Check if the given header has the given value.

fn header_is_ignore_case(&self, name: &str, value: &str) -> bool

Check if the given header has the given value (case-insensitive).