tungstenite::handshake::server

Struct ServerHandshake

pub struct ServerHandshake<S, C> {
    callback: Option<C>,
    config: Option<WebSocketConfig>,
    error_response: Option<ErrorResponse>,
    _marker: PhantomData<S>,
}

Server handshake role.

Fields
callback: Option<C>

Callback which is called whenever the server read the request from the client and is ready to reply to it. The callback returns an optional headers which will be added to the reply which the server sends to the user.

config: Option<WebSocketConfig>

WebSocket configuration.

error_response: Option<ErrorResponse>

Error code/flag. If set, an error will be returned after sending response to the client.

_marker: PhantomData<S>

Internal stream type.

Implementations

impl<S: Read + Write, C: Callback> ServerHandshake<S, C>

pub fn start( stream: S, callback: C, config: Option<WebSocketConfig>, ) -> MidHandshake<Self>

Start server handshake. callback specifies a custom callback which the user can pass to the handshake, this callback will be called when the a websocket client connects to the server, you can specify the callback if you want to add additional header to the client upon join based on the incoming headers.