tungstenite

Module protocol

Generic WebSocket message stream.

Re-exports

pub use self::frame::CloseFrame;

Modules

frame

Utilities to work with raw WebSocket frames.

Structs

WebSocket

WebSocket input-output stream.

WebSocketConfig

The configuration for WebSocket connection.

WebSocketContext

A context for managing WebSocket stream.

Enums

Message

An enum representing the various forms of a WebSocket message.

Role

Indicates a Client or Server role of the websocket