tungstenite::protocol::frame::frame

Enum LengthFormat

enum LengthFormat {
    U8(u8),
    U16,
    U64,
}

Handling of the length format.

Variants

U8(u8)

U16

U64

Implementations

impl LengthFormat

fn for_length(length: u64) -> Self

Get the length format for a given data size.

fn extra_bytes(&self) -> usize

Get the size of the length encoding.

fn length_byte(&self) -> u8

Encode the given length.

fn for_byte(byte: u8) -> Self

Get the length format for a given length byte.