Struct tungstenite::protocol::frame::FrameHeader

pub struct FrameHeader {
    pub is_final: bool,
    pub rsv1: bool,
    pub rsv2: bool,
    pub rsv3: bool,
    pub opcode: OpCode,
    // some fields omitted
}

A struct representing a WebSocket frame header.

Fields
is_final: bool

Indicates that the frame is the last one of a possibly fragmented message.

rsv1: bool

Reserved for protocol extensions.

rsv2: bool

Reserved for protocol extensions.

rsv3: bool

Reserved for protocol extensions.

opcode: OpCode

WebSocket protocol opcode.

Methods

impl FrameHeader

pub fn parse(
cursor: &mut Cursor<impl AsRef<[u8]>>
) -> Result<Option<(Self, u64)>>

Parse a header from an input stream. Returns None if insufficient data and does not consume anything in this case. Payload size is returned along with the header.

pub fn len(&self, length: u64) -> usize

Get the size of the header formatted with given payload length.

pub fn format(
&self,
length: u64,
output: &mut impl Write
) -> Result<()>

Format a header for given payload size.