Struct tungstenite::protocol::frame::FrameSocket

pub struct FrameSocket<Stream> { /* fields omitted */ }

A reader and writer for WebSocket frames.

Methods

impl<Stream> FrameSocket<Stream>

fn new(stream: Stream) -> Self

Create a new frame socket.

fn from_partially_read(stream: Stream, part: Vec<u8>) -> Self

Create a new frame socket from partially read data.

fn into_inner(self) -> (Stream, Vec<u8>)

Extract a stream from the socket.

fn get_ref(&self) -> &Stream

Returns a shared reference to the inner stream.

fn get_mut(&mut self) -> &mut Stream

Returns a mutable reference to the inner stream.

impl<Stream> FrameSocket<Stream> where Stream: Read

fn read_frame(&mut self) -> Result<Option<Frame>>

Read a frame from stream.

impl<Stream> FrameSocket<Stream> where Stream: Write

fn write_frame(&mut self, frame: Frame) -> Result<()>

Write a frame to stream.

This function guarantees that the frame is queued regardless of any errors. There is no need to resend the frame. In order to handle WouldBlock or Incomplete, call write_pending() afterwards.

fn write_pending(&mut self) -> Result<()>

Complete pending write, if any.