Struct tungstenite::protocol::WebSocketConfig

pub struct WebSocketConfig {
    pub max_send_queue: Option<usize>,
    pub write_buffer_size: usize,
    pub max_write_buffer_size: usize,
    pub max_message_size: Option<usize>,
    pub max_frame_size: Option<usize>,
    pub accept_unmasked_frames: bool,
}

The configuration for WebSocket connection.

Fields
max_send_queue: Option<usize>
๐Deprecated

Does nothing, instead use max_write_buffer_size.

write_buffer_size: usize

The target minimum size of the write buffer to reach before writing the data to the underlying stream. The default value is 128 KiB.

If set to 0 each message will be eagerly written to the underlying stream. It is often more optimal to allow them to buffer a little, hence the default value.

Note: flush will always fully write the buffer regardless.

max_write_buffer_size: usize

The max size of the write buffer in bytes. Setting this can provide backpressure in the case the write buffer is filling up due to write errors. The default value is unlimited.

Note: The write buffer only builds up past write_buffer_size when writes to the underlying stream are failing. So the write buffer can not fill up if you are not observing write errors even if not flushing.

Note: Should always be at least write_buffer_size + 1 message and probably a little more depending on error handling strategy.

max_message_size: Option<usize>

The maximum size of an incoming message. None means no size limit. The default value is 64 MiB which should be reasonably big for all normal use-cases but small enough to prevent memory eating by a malicious user.

max_frame_size: Option<usize>

The maximum size of a single incoming message frame. None means no size limit. The limit is for frame payload NOT including the frame header. The default value is 16 MiB which should be reasonably big for all normal use-cases but small enough to prevent memory eating by a malicious user.

accept_unmasked_frames: bool

When set to true, the server will accept and handle unmasked frames from the client. According to the RFC 6455, the server must close the connection to the client in such cases, however it seems like there are some popular libraries that are sending unmasked frames, ignoring the RFC. By default this option is set to false, i.e. according to RFC 6455.