Type Alias tungstenite::ReadBuffer

pub(crate) type ReadBuffer = ReadBuffer<READ_BUFFER_CHUNK_SIZE>;

Aliased Type

struct ReadBuffer {
    storage: Cursor<Vec<u8, Global>>,
    chunk: Box<[u8; 4096], Global>,
}

Fields
storage: Cursor<Vec<u8, Global>>chunk: Box<[u8; 4096], Global>
Implementations

impl<const CHUNK_SIZE: usize> ReadBuffer<CHUNK_SIZE>

pub fn new() -> Self

Create a new empty input buffer.

pub fn with_capacity(capacity: usize) -> Self

Create a new empty input buffer with a given capacity.

pub fn from_partially_read(part: Vec<u8>) -> Self

Create a input buffer filled with previously read data.

pub fn as_cursor(&self) -> &Cursor<Vec<u8>>

Get a cursor to the data storage.

pub fn as_cursor_mut(&mut self) -> &mut Cursor<Vec<u8>>

Get a cursor to the mutable data storage.

pub fn into_vec(self) -> Vec<u8>

Consume the ReadBuffer and get the internal storage.

pub fn read_from<S: Read>(&mut self, stream: &mut S) -> IoResult<usize>

Read next portion of data from the given input stream.

fn clean_up(&mut self)

Cleans ups the part of the vector that has been already read by the cursor.