Crate tungstenite

Lightweight, flexible WebSockets for Rust.

Re-exports

pub use crate::error::Error;

pub use crate::error::Result;

pub use crate::protocol::frame::Utf8Bytes;

pub use crate::protocol::Message;

pub use crate::protocol::WebSocket;

pub use crate::client::client;

pub use crate::client::connect;

pub use crate::client::ClientRequestBuilder;

pub use crate::handshake::client::ClientHandshake;

pub use crate::handshake::server::ServerHandshake;

pub use crate::handshake::HandshakeError;

pub use http;

Modules

buffer

A buffer for reading data from the network.

client

Methods to connect to a WebSocket as a client.

error

Error handling.

handshake

WebSocket handshake control.

protocol

Generic WebSocket message stream.

stream

Convenience wrapper for streams to switch between plain TCP and TLS at runtime.

util

Helper traits to ease non-blocking handling.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

Functions

accept

Accept the given Stream as a WebSocket.

accept_hdr

Accept the given Stream as a WebSocket.

accept_hdr_with_config

Accept the given Stream as a WebSocket.

accept_with_config

Accept the given Stream as a WebSocket.