Crate tungstenite

Lightweight, flexible WebSockets for Rust.

Re-exports

pub use crate::error::Error;

pub use crate::error::Result;

pub use crate::protocol::Message;

pub use crate::protocol::WebSocket;

pub use crate::client::client;

pub use crate::client::connect;

pub use crate::handshake::client::ClientHandshake;

pub use crate::handshake::server::ServerHandshake;

pub use crate::handshake::HandshakeError;

pub use http;

Modules

buffer

A buffer for reading data from the network.

client

Methods to connect to a WebSocket as a client.

error

Error handling.

handshake

WebSocket handshake control.

protocol

Generic WebSocket message stream.

stream

Convenience wrapper for streams to switch between plain TCP and TLS at runtime.

util

Helper traits to ease non-blocking handling.

Enums

Connector

A connector that can be used when establishing connections, allowing to control whether native-tls or rustls is used to create a TLS connection. Or TLS can be disabled with the Plain variant.

Functions

accept

Accept the given Stream as a WebSocket.

accept_hdr

Accept the given Stream as a WebSocket.

accept_hdr_with_config

Accept the given Stream as a WebSocket.

accept_with_config

Accept the given Stream as a WebSocket.

client_tls

Creates a WebSocket handshake from a request and a stream, upgrading the stream to TLS if required.

client_tls_with_config

The same as client_tls() but one can specify a websocket configuration, and an optional connector. If no connector is specified, a default one will be created.