Crate tungstenite

Lightweight, flexible WebSockets for Rust.

Reexports

pub use client::connect;

pub use client::client;

pub use server::accept;

pub use error::Error;

pub use error::Result;

pub use protocol::WebSocket;

pub use handshake::HandshakeError;

pub use handshake::client::ClientHandshake;

pub use handshake::server::ServerHandshake;

Modules

client
Methods to connect to an WebSocket as a client.

error
Error handling.

handshake
WebSocket handshake control.

protocol
Generic WebSocket message stream.

server
Methods to accept an incoming WebSocket connection on a server.

stream
Convenience wrapper for streams to switch between plain TCP and TLS at runtime.

util
Helper traits to ease non-blocking handling.

Enums

Message
An enum representing the various forms of a WebSocket message.