Crate tungstenite

Lightweight, flexible WebSockets for Rust.

Re-exports

pub use http;

pub use crate::client::client;

pub use crate::client::connect;

pub use crate::error::Error;

pub use crate::error::Result;

pub use crate::handshake::client::ClientHandshake;

pub use crate::handshake::server::ServerHandshake;

pub use crate::handshake::HandshakeError;

pub use crate::protocol::WebSocket;

Modules

buffer

A buffer for reading data from the network.

client

Methods to connect to a WebSocket as a client.

error

Error handling.

handshake

WebSocket handshake control.

protocol

Generic WebSocket message stream.

stream

Convenience wrapper for streams to switch between plain TCP and TLS at runtime.

util

Helper traits to ease non-blocking handling.

Enums

Message

An enum representing the various forms of a WebSocket message.

Functions

accept

Accept the given Stream as a WebSocket.

accept_hdr

Accept the given Stream as a WebSocket.

accept_hdr_with_config

Accept the given Stream as a WebSocket.

accept_with_config

Accept the given Stream as a WebSocket.