Crate tungstenite

Lightweight, flexible WebSockets for Rust.

Re-exports

pub use http;

pub use crate::client::client;

pub use crate::client::connect;

pub use crate::error::Error;

pub use crate::error::Result;

pub use crate::handshake::client::ClientHandshake;

pub use crate::handshake::server::ServerHandshake;

pub use crate::handshake::HandshakeError;

pub use crate::protocol::WebSocket;

pub use crate::server::accept;

pub use crate::server::accept_hdr;

Modules

client

Methods to connect to a WebSocket as a client.

error

Error handling.

handshake

WebSocket handshake control.

protocol

Generic WebSocket message stream.

server

Methods to accept an incoming WebSocket connection on a server.

stream

Convenience wrapper for streams to switch between plain TCP and TLS at runtime.

util

Helper traits to ease non-blocking handling.

Enums

Message

An enum representing the various forms of a WebSocket message.