Module tungstenite::client

Methods to connect to a WebSocket as a client.

Traits

IntoClientRequest

Trait for converting various types into HTTP requests used for a client connection.

Functions

client

Do the client handshake over the given stream.

client_with_config

Do the client handshake over the given stream given a web socket configuration. Passing None as configuration is equal to calling client() function.

connect

Connect to the given WebSocket in blocking mode.

connect_with_config

Connect to the given WebSocket in blocking mode.

uri_mode

Get the mode of the given URL.