Function tungstenite::client::client_with_config

pub fn client_with_config<Stream, Req>(
    request: Req,
    stream: Stream,
    config: Option<WebSocketConfig>,
) -> Result<(WebSocket<Stream>, Response), HandshakeError<ClientHandshake<Stream>>>
where
    Stream: Read + Write,
    Req: IntoClientRequest,

Do the client handshake over the given stream given a web socket configuration. Passing None as configuration is equal to calling client() function.

Use this function if you need a nonblocking handshake support or if you want to use a custom stream like mio::net::TcpStream or openssl::ssl::SslStream. Any stream supporting Read + Write will do.