Trait tungstenite::client::IntoClientRequest

pub trait IntoClientRequest {
    fn into_client_request(self) -> Result<Request>;
}

Trait for converting various types into HTTP requests used for a client connection.

This trait is implemented by default for string slices, strings, url::Url, http::Uri and http::Request<()>.

Required methods

fn into_client_request(self) -> Result<Request>

Convert into a Request that can be used for a client connection.