Struct tungstenite::handshake::headers::Headers

pub struct Headers { /* fields omitted */ }

HTTP request or response headers.

Methods

impl Headers

pub fn find_first(&self, name: &str) -> Option<&[u8]>

Get first header with the given name, if any.

pub fn find<'headers, 'name>(
&'headers self,
name: &'name str
) -> HeadersIter<'name, 'headers>

Iterate over all headers with the given name.

pub fn header_is(&self, name: &str, value: &str) -> bool

Check if the given header has the given value.

pub fn header_is_ignore_case(&self, name: &str, value: &str) -> bool

Check if the given header has the given value (case-insensitive).

pub fn iter(&self) -> Iter<(String, Box<[u8]>)>

Allows to iterate over available headers.