Struct tungstenite::handshake::server::ServerHandshake

pub struct ServerHandshake<S, C> { /* fields omitted */ }

Server handshake role.

Methods

impl<S: Read + Write, C: Callback> ServerHandshake<S, C>

pub fn start(stream: S, callback: C) -> MidHandshake<Self>

Start server handshake. callback specifies a custom callback which the user can pass to the handshake, this callback will be called when the a websocket client connnects to the server, you can specify the callback if you want to add additional header to the client upon join based on the incoming headers.