Struct tungstenite::protocol::frame::FrameSocket

pub struct FrameSocket<Stream> { /* private fields */ }

A reader and writer for WebSocket frames.

Implementations

impl<Stream> FrameSocket<Stream>

pub fn new(stream: Stream) -> Self

Create a new frame socket.

pub fn from_partially_read(stream: Stream, part: Vec<u8>) -> Self

Create a new frame socket from partially read data.

pub fn into_inner(self) -> (Stream, Vec<u8>)

Extract a stream from the socket.

pub fn get_ref(&self) -> &Stream

Returns a shared reference to the inner stream.

pub fn get_mut(&mut self) -> &mut Stream

Returns a mutable reference to the inner stream.

impl<Stream> FrameSocket<Stream> where
Stream: Read,

pub fn read_frame(&mut self, max_size: Option<usize>) -> Result<Option<Frame>>

Read a frame from stream.

impl<Stream> FrameSocket<Stream> where
Stream: Write,

pub fn write_frame(&mut self, frame: Frame) -> Result<()>

Write a frame to stream.

This function guarantees that the frame is queued regardless of any errors. There is no need to resend the frame. In order to handle WouldBlock or Incomplete, call write_pending() afterwards.

pub fn write_pending(&mut self) -> Result<()>

Complete pending write, if any.