Struct tungstenite::protocol::WebSocket

pub struct WebSocket<Stream> { /* fields omitted */ }

WebSocket input-output stream.

This is THE structure you want to create to be able to speak the WebSocket protocol. It may be created by calling connect, accept or client functions.

Methods

impl<Stream> WebSocket<Stream>

pub fn from_raw_socket(stream: Stream, role: Role) -> Self

Convert a raw socket into a WebSocket without performing a handshake.

pub fn from_partially_read(stream: Stream, part: Vec<u8>, role: Role) -> Self

Convert a raw socket into a WebSocket without performing a handshake.

pub fn get_ref(&self) -> &Stream

Returns a shared reference to the inner stream.

pub fn get_mut(&mut self) -> &mut Stream

Returns a mutable reference to the inner stream.

impl<Stream: Read + Write> WebSocket<Stream>

pub fn read_message(&mut self) -> Result<Message>

Read a message from stream, if possible.

This function sends pong and close responses automatically. However, it never blocks on write.

pub fn write_message(&mut self, message: Message) -> Result<()>

Send a message to stream, if possible.

This function guarantees that the frame is queued regardless of any errors. There is no need to resend the frame. In order to handle WouldBlock or Incomplete, call write_pending() afterwards.

Note that only the last pong frame is stored to be sent, and only the most recent pong frame is sent if multiple pong frames are queued up.

pub fn close(&mut self, code: Option<CloseFrame>) -> Result<()>

Close the connection.

This function guarantees that the close frame will be queued. There is no need to call it again, just like write_message().

pub fn write_pending(&mut self) -> Result<()>

Flush the pending send queue.