Module tungstenite::tls

Connection helper.

Modules

encryption 🔒

Enums

Connector

A connector that can be used when establishing connections, allowing to control whether native-tls or rustls is used to create a TLS connection. Or TLS can be disabled with the Plain variant.

Functions

client_tls

Creates a WebSocket handshake from a request and a stream, upgrading the stream to TLS if required.

client_tls_with_config

The same as client_tls() but one can specify a websocket configuration, and an optional connector. If no connector is specified, a default one will be created.

Type Aliases

TlsHandshakeError 🔒